caribbean-poker/game-master/requirements.txt
pyyaml
pywaves
rsa
pycryptodome